requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pydantic>=2.9.0
pymongo==4.6.0
requests==2.31.0
email-validator==2.1.0
google-api-python-client==2.149.0
google-auth==2.35.0
google-auth-oauthlib==1.2.1
python-dateutil==2.9.0.post0
pytz==2024.1
twilio==9.2.3
APScheduler==3.10.4
